% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_push_local_dockerfile.R
\name{pull_push_local_dockerfile}
\alias{pull_push_local_dockerfile}
\title{Pull or push image from dockerhub
Assumes your built image is named after your dockerhub username}
\usage{
pull_push_local_dockerfile(pull_push, dockerhub_username, project_name)
}
\arguments{
\item{pull_push}{Specify 'pull' or 'push'}

\item{dockerhub_username}{username for dockerhub}

\item{project_name}{built image name}
}
\description{
Pull or push image from dockerhub
Assumes your built image is named after your dockerhub username
}
\examples{
pull_push_local_dockerfile('push', 'my_username', 'my_project')
}
