% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_local_dockerfile.R
\name{run_local_dockerfile}
\alias{run_local_dockerfile}
\title{Run a built dockerfile locally, accessed through the 8787 port
Assumes your built image is named after your dockerhub username}
\usage{
run_local_dockerfile(dockerhub_username, project_name)
}
\arguments{
\item{dockerhub_username}{username for dockerhub}

\item{project_name}{built image name}
}
\value{
Opens url with container running
}
\description{
Run a built dockerfile locally, accessed through the 8787 port
Assumes your built image is named after your dockerhub username
}
\examples{
run_local_dockerfile('my_username', 'my_project')
}
