% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_binder_dockerfile.R
\name{write_binder_dockerfile}
\alias{write_binder_dockerfile}
\title{Writes dockerfile to .binder/ to be used to run a binder container
Assumes your built image is named after your dockerhub username,
and has been pushed to dockerhub}
\usage{
write_binder_dockerfile(dockerhub_username, project_name,
  project_maintainer, maintainer_email)
}
\arguments{
\item{dockerhub_username}{Username for dockerhub}

\item{project_name}{Built image name}

\item{project_maintainer}{Name of maintainer}

\item{maintainer_email}{Email address of maintainer}
}
\value{
Creates .binder folder and writes binder-enabled Dockerfile to it
}
\description{
Writes dockerfile to .binder/ to be used to run a binder container
Assumes your built image is named after your dockerhub username,
and has been pushed to dockerhub
}
\examples{
write_binder_dockerfile('my_username', 'my_project', 'my_name', 'my_email')
}
