% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_greta_dockerfile.R
\name{write_greta_dockerfile}
\alias{write_greta_dockerfile}
\title{Writes dockerfile with greta installed}
\usage{
write_greta_dockerfile(
  maintainer_name,
  maintainer_email,
  r_ver = NULL,
  date = NULL
)
}
\arguments{
\item{maintainer_name}{Name of maintainer}

\item{maintainer_email}{Email address of maintainer}

\item{r_ver}{optional to specify R version of base image}

\item{date}{optional to specify date for R package version installs}
}
\value{
Writes Dockerfile to current directory
}
\description{
Writes dockerfile with greta installed
}
\examples{
{
write_greta_dockerfile('my_name', 'my_email', r_ver = '4.3.0',
date = '2023-04-12')
}
}
