% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_local_dockerfile.R
\name{write_local_dockerfile}
\alias{write_local_dockerfile}
\title{Writes dockerfile}
\usage{
write_local_dockerfile(maintainer_name, maintainer_email, r_ver = NULL)
}
\arguments{
\item{maintainer_name}{Name of maintainer}

\item{maintainer_email}{Email address of maintainer}

\item{r_ver}{optional to specify R version of base image}
}
\value{
Writes Dockerfile to current directory
}
\description{
Writes dockerfile
}
\examples{
write_local_dockerfile('my_name', 'my_email', '3.5.0')
}
