% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_local_dockerfile.R
\name{build_local_dockerfile}
\alias{build_local_dockerfile}
\title{Docker build local image
Assumes your built image is named after your dockerhub username}
\usage{
build_local_dockerfile(dockerhub_username, project_name)
}
\arguments{
\item{dockerhub_username}{username for dockerhub}

\item{project_name}{built image name}
}
\description{
Docker build local image
Assumes your built image is named after your dockerhub username
}
\examples{
build_local_dockerfile('my_username', 'my_project_name')
}
